fix: add throwing error in case file exist to create func

+++ src/fs/create.js
@@ -1,7 +1,11 @@
 import { writeFile } from 'node:fs/promises';
+import { existsSync } from 'node:fs';
 
 const create = async () => {
     try {
+        if (existsSync("files/fresh.txt")) {
+            throw new Error('FS operation failed: File already exists');
+        }
         const content = 'I am fresh and young';
         await writeFile('files/fresh.txt', content);
     } catch (err) {
@@ -11,9 +15,6 @@ const create = async () => {
 
 await create();
 
-/*
-create.js - 
-implement function that creates new file fresh.txt with content 
-'I am fresh and young' inside of the files folder
+/*implement function that creates new file fresh.txt with content 'I am fresh and young' inside of the files folder
 (if file already exists Error with message FS operation failed must be thrown)
 */
